backend/requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# AI and ML Integration
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
langchain-openai==0.0.2
langchain-anthropic==0.0.1

# Voice Processing
azure-cognitiveservices-speech==1.34.0
google-cloud-speech==2.21.0
pydub==0.25.1
librosa==0.10.1

# WebSocket and Real-time
socketio==5.10.0
python-socketio[asyncio_client]==5.10.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Email and Notifications
sendgrid==6.10.0
jinja2==3.1.2

# File Processing and Storage
boto3==1.34.0
pillow==10.1.0
python-magic==0.4.27

# Security and Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Monitoring and Logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Calendar Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Microsoft Graph API
msal==1.24.1
requests==2.31.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6
